buildvgl -v: Fail if subscript fails
If -v was passed to buildvgl (which our CI system always does), then the
exit status of buildvgl was the exit status of tee, not the exit status
of buildvgl.$PLATFORM.  Effectively, that meant that errors in the
official build jobs were not communicated to GitHub Actions.
'set -o pipefail' causes the exit status of buildvgl to be the logical
AND of the exit statuses of buildvgl.$PLATFORM and tee.